feat(ServicoDao): adicionar metodo findAll

+++ src/modulo-05-java/dia-01/src/br/com/cwi/crescer/model/Servico.java
@@ -10,7 +10,7 @@ public class Servico {
         return idServico;
     }
 
-    public void setIdServico(int idServico) {
+    public void setIdServico(long idServico) {
         this.idServico = idServico;
     }
